+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Correlation matrix
 - ECDF and complementary ECDF on stats tab
 - Hypothesis test for normality improved
+- Learning curve in training phase for OLS model
 
 ## [2.2.0] - 2026/01/28
 -Stats tab